Fix sample (#296)
* [FIX] fix bugs with building sample; add scripts for building on linux (#292)


* change sample name

* 统一命名规范

* 统一android头文件规范
