+++ examples/android/demo/src/main/jni/cc/face_detector_jni.cc
@@ -1,8 +1,19 @@
+// Tencent is pleased to support the open source community by making TNN available.
 //
-// Created by tencent on 2020-04-29.
+// Copyright (C) 2020 THL A29 Limited, a Tencent company. All rights reserved.
 //
-#include "facedetector_jni.h"
-#include "UltraFaceDetector.h"
+// Licensed under the BSD 3-Clause License (the "License"); you may not use this file except
+// in compliance with the License. You may obtain a copy of the License at
+//
+// https://opensource.org/licenses/BSD-3-Clause
+//
+// Unless required by applicable law or agreed to in writing, software distributed
+// under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+// CONDITIONS OF ANY KIND, either express or implied. See the License for the
+// specific language governing permissions and limitations under the License.
+
+#include "face_detector_jni.h"
+#include "ultra_face_detector.h"
 #include "kannarotate.h"
 #include "yuv420sp_to_rgb_fast_asm.h"
 #include <jni.h>
