+++ examples/android/demo/src/main/jni/cc/face_detector_jni.h
@@ -12,29 +12,21 @@
 // CONDITIONS OF ANY KIND, either express or implied. See the License for the
 // specific language governing permissions and limitations under the License.
 
-#ifndef ANDROID_TNN_JNI_H_
-#define ANDROID_TNN_JNI_H_
+#ifndef ANDROID_FACE_DETECTOR_JNI_H_
+#define ANDROID_FACE_DETECTOR_JNI_H_
 
 #include <jni.h>
-#include <string>
-#include <android/log.h>
-
+#define TNN_FACE_DETECTOR(sig) Java_com_tencent_tnn_demo_FaceDetector_##sig
 #ifdef __cplusplus
 extern "C" {
 #endif
-
-JNIEXPORT JNICALL jint Java_com_tencent_tnn_demo_TNNLib_init(
-        JNIEnv *env, jobject thiz, jstring protoFilePath, jstring modelFilePath, jstring device);
-
-JNIEXPORT JNICALL jfloatArray Java_com_tencent_tnn_demo_TNNLib_forward(
-        JNIEnv *env, jobject thiz, jobject imageSource);
-
-JNIEXPORT JNICALL jint Java_com_tencent_tnn_demo_TNNLib_deinit(
-        JNIEnv *env, jobject thiz);
-
+JNIEXPORT JNICALL jint TNN_FACE_DETECTOR(init)(JNIEnv *env, jobject thiz, jstring modelPath, jint width, jint height, jfloat scoreThreshold, jfloat iouThreshold, jint topk, jint computUnitType);
+JNIEXPORT JNICALL jint TNN_FACE_DETECTOR(deinit)(JNIEnv *env, jobject thiz);
+JNIEXPORT JNICALL jobjectArray TNN_FACE_DETECTOR(detectFromStream)(JNIEnv *env, jobject thiz, jbyteArray yuv420sp, jint width, jint height, jint rotate);
+JNIEXPORT JNICALL jobjectArray TNN_FACE_DETECTOR(detectFromImage)(JNIEnv *env, jobject thiz, jobject imageSource, jint width, jint height);
 
 #ifdef __cplusplus
 }
 #endif
 
-#endif // ANDROID_TNN_JNI_H_
+#endif // ANDROID_FACE_DETECTOR_JNI_H_
