+++ examples/android/demo/src/main/jni/cc/helper_jni.h
@@ -1,9 +1,20 @@
+// Tencent is pleased to support the open source community by making TNN available.
 //
-// Created by tencent on 2020-04-30.
+// Copyright (C) 2020 THL A29 Limited, a Tencent company. All rights reserved.
 //
+// Licensed under the BSD 3-Clause License (the "License"); you may not use this file except
+// in compliance with the License. You may obtain a copy of the License at
+//
+// https://opensource.org/licenses/BSD-3-Clause
+//
+// Unless required by applicable law or agreed to in writing, software distributed
+// under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+// CONDITIONS OF ANY KIND, either express or implied. See the License for the
+// specific language governing permissions and limitations under the License.
 
-#ifndef ANDROID_HELPER_JNI_H
+#ifndef ANDROID_HELPER_JNI_H_
 #define ANDROID_HELPER_JNI_H
+
 #include <string>
 #include <jni.h>
 #define TNN_HELPER(sig) Java_com_tencent_tnn_demo_Helper_##sig
@@ -18,4 +29,5 @@ JNIEXPORT JNICALL jstring TNN_HELPER(getBenchResult)(JNIEnv *env, jobject thiz);
 #ifdef __cplusplus
 }
 #endif
-#endif //ANDROID_HELPER_JNI_H
+
+#endif // ANDROID_HELPER_JNI_H_
