+++ examples/base/image_classifier.cc
@@ -12,8 +12,8 @@
 // CONDITIONS OF ANY KIND, either express or implied. See the License for the
 // specific language governing permissions and limitations under the License.
 
-#include "ImageClassifier.h"
-#include "examples/samples/timer.h"
+#include "image_classifier.h"
+#include "sample_timer.h"
 #include <cmath>
 
 ImageClassifier::ImageClassifier() {}
