+++ examples/ios/TNNExamples/TNNCameraPreviewController/TNNCameraPreviewController.mm
@@ -6,7 +6,7 @@
 #import <CoreMedia/CoreMedia.h>
 #import <tnn/tnn.h>
 #import "UIImage+Utility.h"
-#import "UltraFaceDetector.h"
+#import "ultra_face_detector.h"
 
 using namespace std;
 using namespace TNN_NS;
