+++ examples/linux/include/utils.h
